+++ src/components/index.js
@@ -1,0 +1,33 @@
+// For imports and exports
+
+import { Home } from "./Home";
+import { PuffLoader } from "./Loader";
+import { Navbar } from "./Navbar";
+import { ProductCard } from "./Products/Productcard";
+import { ProductDetail } from "./Products/ProductDetail";
+import { Sidebar } from "./Products/Sidebar";
+import { Wishcard } from "./Wishlist/Wishcard";
+import { Wishlist } from "./Wishlist/Wishlist";
+import { Wishlistempty } from "./Wishlist/Wishlistempty";
+import { Cart } from "./Cart/Cart";
+import { CartCard } from "./Cart/CartCard";
+import { CartEmpty } from "./Cart/CartEmpty";
+import { Products } from "./Products/Products";
+import { Modal } from "./Modal";
+
+export {
+  Home,
+  PuffLoader,
+  Navbar,
+  ProductCard,
+  ProductDetail,
+  Sidebar,
+  Wishcard,
+  Wishlist,
+  Cart,
+  CartCard,
+  Wishlistempty,
+  CartEmpty,
+  Products,
+  Modal,
+};
