+++ src/hooks/index.js
@@ -1,0 +1,13 @@
+import { useCartData } from "./useCart/useCartData";
+import { useLocalStorage } from "./useLocalStorage";
+import { useProductData } from "./useProductData/useProductData";
+import { useProductDetails } from "./useProductDetails";
+import { useWishlistData } from "./useWishlist/useWishlistData";
+
+export {
+  useCartData,
+  useLocalStorage,
+  useProductData,
+  useProductDetails,
+  useWishlistData,
+};
